requirements.txt
# TAMER OCR - Requirements
# =======================

# Core
torch>=2.0.0
torchvision>=0.15.0

# Model
timm>=0.9.0

# Data Processing
opencv-python-headless>=4.7.0
Pillow>=9.0.0
albumentations>=1.3.0
numpy>=1.23.0

# Datasets & Downloads
datasets>=2.12.0
huggingface_hub>=0.16.0
requests>=2.28.0

# Logging & Utilities
tqdm>=4.65.0

# Optional (for Kaggle downloads)
# kaggle>=1.5.0
